refactor(elo): move Bayesian forecast functions out of forecast.py
Avoids importing arviz at module load time so non-Bayesian commands
(snapshot, run, evaluate) no longer fail when the optional [bayesian]
extras are not installed.
